refactor: Result 객체의 이름을 LadderGame으로 변경한다
